+++ src/practicalli/application.clj
@@ -38,4 +38,7 @@
   Return
   - new template.edn configuration"
   [edn data]
+
+  ;; Link to Practicalli Clojure Project Templates guide
+  (println "Template guide: https://practical.li/clojure/clojure-cli/projects/templates/")
   edn)